Add GHA workflow name to each example (fix for #223) (#222)
* Add GHA workflow name to each example

The text of the README doesn't actually specify what to pass as the `name` parameter of `usethis::use_github_action()` for each option. It's not hard to find out, or to guess, but I think the documentation would be better if this information were explicitly included. I have omitted Bioconductor as the code for this is already included.

* Amend use_github_action() guidance

The guidance now just lists the actual command you would use.