Move the gfortran gcov executable
Rather than prepending /usr/bin, which breaks other workflows which put
tools before /usr/bin, such as setup-python we instead rename the gcov
executable shipped with gfortran. We generally don't want to use the
gcov executable in the gfortran binary, as this is a gcc gcov, rather we
want to use the gcov shipped with llvm. Renaming accomplishes this.

Fixes #250