Include a wildcard in the path
This allows the exclusion to work. The exclusions are only applied to the list of files, and directories are not recursed into.
So without the wildcard the exclusion will never match. However with a single wildcard it will and do what we want.

Fixes https://github.com/r-lib/actions/issues/264#issuecomment-806373311