[setup-r-dependencies]: restore renv compatibility
It seems that renv will set `.Library.site` aside, so people cannot
call pak if renv is active. This commit tries to fix the renv
compatibility, by installing pak into the default library if
renv is active.

We have a new env var now, `R_LIB_FOR_PAK`, which is the library where
pak is installed. This is set early on, and then used throughout.
It is set to `R_LIBS_SITE` for non-renv projects, and `.libPaths()[1]`
if renv is active.